Use Postgresql 9.6 on Ubuntu 16.04
I think its nice to use newest Postgresql which is version 9.6 since it has parallel query ability. But since the binary is not available yet in Ubuntu repository (but it is official already in Postgresql repository), the repository to add is unique to each Ubuntu version. This PR only assume you are using Ubuntu 16.04.